#165 Refactor code out of Template
